+++ fields.html
@@ -12,7 +12,7 @@
         <input id="submit" type="submit" name="submit" />
       </div>
     </form>
-    <script type="text/javascript" src="/dist/securesubmit.min.js"></script>
+    <script type="text/javascript" src="/dist/securesubmit.js"></script>
     <script type="text/javascript">
       (function () {
         window.hps = new HPS({
