+++ src/types/CardType.ts
@@ -1,0 +1,6 @@
+module Heartland {
+  export interface CardType {
+      code: string;
+      regex: RegExp;
+  }
+}
