+++ tsconfig.json
@@ -5,5 +5,5 @@
         "sourceMap": true,
         "noImplicitAny": true
     },
-    "files": [ "src/securesubmit" ]
+    "files": [ "src/SecureSubmit" ]
 }
